--- src/components/AddMileage/index.ts
@@ -1,10 +1,0 @@
-export { default as AddMileageModal } from './AddMileageModal';
-export { default as DeleteSubmittedMileageModal } from './DeleteSubmittedMileageModal';
-export { default as EditSubmittedMileageModal } from './EditSubmittedMileageModal';
-export { default as EtcMileageSection } from './EtcMileageSection';
-export { default as EtcMileageTable } from './EtcMileageTable';
-export { default as FileDownloadButton } from './FileDownloadButton';
-export { default as GuideDescSection } from './GuideDescSection';
-export { default as SubmittedMileageModal } from './SubmittedMileageModal';
-export { default as SubmittedMileageSection } from './SubmittedMileageSection';
-export { default as SubmittedMileageTable } from './SubmittedMileageTable';
