+++ src/hooks/index.ts
@@ -1,8 +1,6 @@
-export { default as useEditMileageForm } from './useEditMileageForm';
 export { default as useFile } from './useFile';
 export { default as useFileWithType } from './useFileWithType';
 export { default as useInput } from './useInput';
 export { default as useInputWithValidate } from './useInputWithValidate';
 export { default as useLogin } from './useLogin';
-export { default as useNewMileageForm } from './useNewMileageForm';
 export { default as useOpenModal } from './useOpenModal';
