+++ src/pages/AddMileagePage/components/EditSubmittedMileageModal.tsx
@@ -1,31 +1,50 @@
-import { Button, Dropdown, Flex, FormField, Modal, Text } from '@/components';
-import { FileDownloadButton, GuideDescSection } from '@/components/AddMileage';
+import { EditIcon } from '@/assets';
+import {
+  Button,
+  Dropdown,
+  Flex,
+  FormField,
+  Modal,
+  Text,
+  UploadButton,
+} from '@/components';
 import { MAX_RESPONSIVE_WIDTH } from '@/constants/system';
 import { useOpenModal } from '@/hooks';
-import { trackSubmittedMileageModalButton } from '@/service/amplitude/trackEvent';
-import { SubmittedMileageResponse } from '@/types/mileage';
+import { trackSubmittedMileageModalEditButton } from '@/service/amplitude/trackEvent';
 import { styled, useMediaQuery, useTheme } from '@mui/material';
+
+import { useEditMileageForm } from '../hooks/useEditMileageForm';
+import { SubmittedMileageResponse } from '../types/addMileage';
+import { GuideDescSection } from './GuideDescSection';
 
 interface Props {
   item: SubmittedMileageResponse;
 }
 
-const SubmittedMileageModal = ({ item }: Props) => {
+export const EditSubmittedMileageModal = ({ item }: Props) => {
   const isMobile = useMediaQuery(MAX_RESPONSIVE_WIDTH);
   const theme = useTheme();
   const { open, toggleModal } = useOpenModal();
+
+  const { desc1, desc2, file, handleSubmit } = useEditMileageForm({
+    item,
+    toggleModal,
+  });
+
+  const handleSubmitForm = (
+    e: React.FormEvent<HTMLFormElement> | React.MouseEvent<HTMLButtonElement>,
+  ) => {
+    handleSubmit(e);
+  };
 
   return (
     <Modal
       open={open}
       toggleModal={toggleModal}
       trigger={
-        <Button
-          label="상세보기"
-          isRound
-          style={{ width: '100px' }}
-          onClick={() => trackSubmittedMileageModalButton()}
-        />
+        <S.IconButton onClick={() => trackSubmittedMileageModalEditButton()}>
+          <EditIcon />
+        </S.IconButton>
       }
       size="large"
       hasCloseButton
@@ -33,14 +52,14 @@ const SubmittedMileageModal = ({ item }: Props) => {
         backgroundColor: theme.palette.background.default,
       }}
     >
-      <Modal.Header>마일리지 등록 항목 상세보기</Modal.Header>
+      <Modal.Header>마일리지 등록 항목 수정하기</Modal.Header>
       <Modal.Body
         position="center"
         style={{ width: isMobile ? '100%' : '85%', margin: '2rem auto' }}
       >
         <GuideDescSection />
 
-        <S.Form>
+        <S.Form onSubmit={handleSubmitForm}>
           <FormField
             direction={isMobile ? 'column' : 'row'}
             style={{
@@ -62,7 +81,6 @@ const SubmittedMileageModal = ({ item }: Props) => {
                 selectedItem={item.semester}
                 setSelectedItem={() => {}}
                 width="100%"
-                disabled
               />
               <FormField.Box />
             </Flex.Column>
@@ -82,8 +100,8 @@ const SubmittedMileageModal = ({ item }: Props) => {
               <FormField.Input
                 placeholder={'활동 항목에 대해 작성 해주세요'}
                 fullWidth
-                value={item.description1}
-                disabled
+                value={desc1.value}
+                onChange={desc1.handleChange}
               />
             </Flex.Column>
           </FormField>
@@ -101,8 +119,8 @@ const SubmittedMileageModal = ({ item }: Props) => {
               <FormField.Input
                 placeholder="활동을 자세히 설명해주세요"
                 fullWidth
-                value={item.description2}
-                disabled
+                value={desc2.value}
+                onChange={desc2.handleChange}
                 minRows={4}
                 maxRows={4}
                 multiline
@@ -120,14 +138,29 @@ const SubmittedMileageModal = ({ item }: Props) => {
                 ...theme.typography.body2,
               }}
             />
-            <Flex.Row gap="1rem" align="center">
+            <Flex.Row gap="1rem" align="center" wrap="wrap">
+              <UploadButton
+                label="첨부파일 업로드"
+                onUpload={file.handleChange}
+              />
               <Flex.Column>
-                {item.file ? (
-                  <FileDownloadButton item={item} />
-                ) : (
-                  <Text style={{ ...theme.typography.body2 }}>
-                    제출된 파일이 없습니다.
+                {item.file || file.value ? (
+                  <Text
+                    style={{
+                      ...theme.typography.body2,
+                      whiteSpace: 'nowrap',
+                      overflow: 'hidden',
+                      textOverflow: 'ellipsis',
+                    }}
+                  >
+                    {file.value ? file.value.name : item.file}
                   </Text>
+                ) : (
+                  <>
+                    <Text style={{ ...theme.typography.body2 }}>
+                      제출된 파일이 없습니다.
+                    </Text>
+                  </>
                 )}
               </Flex.Column>
             </Flex.Row>
@@ -135,15 +168,19 @@ const SubmittedMileageModal = ({ item }: Props) => {
           </FormField>
 
           <Flex.Row justify="center" gap="2rem" margin="2rem 0 0">
-            <S.CloseButton label="닫기" onClick={toggleModal} size="large" />
+            <S.CloseButton
+              label="닫기"
+              onClick={toggleModal}
+              size="large"
+              color="grey"
+            />
+            <S.SubmitButton type="submit" label="저장하기" size="large" />
           </Flex.Row>
         </S.Form>
       </Modal.Body>
     </Modal>
   );
 };
-
-export default SubmittedMileageModal;
 
 const S = {
   Form: styled('form')`
@@ -153,7 +190,26 @@ const S = {
     padding-top: 2rem;
     width: 100%;
   `,
-  CloseButton: styled(Button)`
+  IconButton: styled('button')`
+    align-items: center;
+    background-color: ${({ theme }) => theme.palette.grey100};
+    border: 2px solid ${({ theme }) => theme.palette.grey200};
+    border-radius: 0.5rem;
+    display: flex;
+    height: 30px;
+    justify-content: center;
+    width: 30px;
+  `,
+  SubmitButton: styled(Button)`
     width: 300px;
+  `,
+  CloseButton: styled(Button)`
+    background-color: ${({ theme }) => theme.palette.grey300};
+    width: 300px;
+
+    &:hover,
+    &:active {
+      background-color: ${({ theme }) => theme.palette.grey400};
+    }
   `,
 };
