+++ src/pages/AddMileagePage/components/GuideDescSection.tsx
@@ -4,13 +4,9 @@ import { MAX_RESPONSIVE_WIDTH } from '@/constants/system';
 import { boxShadow } from '@/styles/common';
 import { styled, useMediaQuery, useTheme } from '@mui/material';
 
-const guides = [
-  'SW 관련 자격증 또는 개인별 추가 실적 제출',
-  '첨부파일은 pdf만 업로드 가능',
-  '업로드 할 첨부파일 : 자격증 사본 또는 실적을 증빙할 수 있는 서류 사본',
-];
+import { infoGuide } from '../constants/infoGuide';
 
-const GuideDescSection = () => {
+export const GuideDescSection = () => {
   const theme = useTheme();
   const isMobile = useMediaQuery(MAX_RESPONSIVE_WIDTH);
 
@@ -30,7 +26,7 @@ const GuideDescSection = () => {
       </Flex.Column>
 
       <S.GuideWrapper padding=".5rem 1rem">
-        {guides.map(guide => (
+        {infoGuide.map(guide => (
           <Flex.Row align="center" gap="0.25rem" key={guide}>
             <CheckIcon />
             <Text
@@ -46,8 +42,6 @@ const GuideDescSection = () => {
     </S.RowSection>
   );
 };
-
-export default GuideDescSection;
 
 const S = {
   RowSection: styled(Flex.Row)`
