+++ src/pages/AddMileagePage/components/SubmittedMileageModal.tsx
@@ -1,42 +1,21 @@
-import {
-  Button,
-  Dropdown,
-  Flex,
-  FormField,
-  Modal,
-  Text,
-  UploadButton,
-} from '@/components';
-import { GuideDescSection } from '@/components/AddMileage';
+import { Button, Dropdown, Flex, FormField, Modal, Text } from '@/components';
 import { MAX_RESPONSIVE_WIDTH } from '@/constants/system';
-import { useNewMileageForm, useOpenModal } from '@/hooks';
-import {
-  trackAddNewMileageButton,
-  trackAddNewMileageModalButton,
-} from '@/service/amplitude/trackEvent';
+import { useOpenModal } from '@/hooks';
+import { trackSubmittedMileageModalButton } from '@/service/amplitude/trackEvent';
 import { styled, useMediaQuery, useTheme } from '@mui/material';
 
+import { SubmittedMileageResponse } from '../types/addMileage';
+import { FileDownloadButton } from './FileDownloadButton';
+import { GuideDescSection } from './GuideDescSection';
+
 interface Props {
-  semester: string;
-  subitemId: number;
+  item: SubmittedMileageResponse;
 }
 
-const AddMileageModal = ({ semester, subitemId }: Props) => {
+export const SubmittedMileageModal = ({ item }: Props) => {
   const isMobile = useMediaQuery(MAX_RESPONSIVE_WIDTH);
   const theme = useTheme();
   const { open, toggleModal } = useOpenModal();
-
-  const { desc1, desc2, file, handleSubmit } = useNewMileageForm(
-    semester,
-    subitemId,
-    toggleModal,
-  );
-
-  const handleSubmitForm = (
-    e: React.FormEvent<HTMLFormElement> | React.MouseEvent<HTMLButtonElement>,
-  ) => {
-    handleSubmit(e);
-  };
 
   return (
     <Modal
@@ -44,10 +23,10 @@ const AddMileageModal = ({ semester, subitemId }: Props) => {
       toggleModal={toggleModal}
       trigger={
         <Button
-          label="등록하기"
+          label="상세보기"
           isRound
           style={{ width: '100px' }}
-          onClick={() => trackAddNewMileageModalButton()}
+          onClick={() => trackSubmittedMileageModalButton()}
         />
       }
       size="large"
@@ -56,14 +35,14 @@ const AddMileageModal = ({ semester, subitemId }: Props) => {
         backgroundColor: theme.palette.background.default,
       }}
     >
-      <Modal.Header>마일리지 활동 등록하기</Modal.Header>
+      <Modal.Header>마일리지 등록 항목 상세보기</Modal.Header>
       <Modal.Body
         position="center"
         style={{ width: isMobile ? '100%' : '85%', margin: '2rem auto' }}
       >
         <GuideDescSection />
 
-        <S.Form onSubmit={handleSubmitForm}>
+        <S.Form>
           <FormField
             direction={isMobile ? 'column' : 'row'}
             style={{
@@ -81,10 +60,11 @@ const AddMileageModal = ({ semester, subitemId }: Props) => {
             />
             <Flex.Column width="100%">
               <Dropdown
-                items={[semester]}
-                selectedItem={semester}
+                items={[item.semester]}
+                selectedItem={item.semester}
                 setSelectedItem={() => {}}
                 width="100%"
+                disabled
               />
               <FormField.Box />
             </Flex.Column>
@@ -104,10 +84,9 @@ const AddMileageModal = ({ semester, subitemId }: Props) => {
               <FormField.Input
                 placeholder={'활동 항목에 대해 작성 해주세요'}
                 fullWidth
-                value={desc1.value}
-                onChange={desc1.handleChange}
+                value={item.description1}
+                disabled
               />
-              <FormField.ErrorMessage value={desc1.errorMessage} />
             </Flex.Column>
           </FormField>
 
@@ -124,8 +103,8 @@ const AddMileageModal = ({ semester, subitemId }: Props) => {
               <FormField.Input
                 placeholder="활동을 자세히 설명해주세요"
                 fullWidth
-                value={desc2.value}
-                onChange={desc2.handleChange}
+                value={item.description2}
+                disabled
                 minRows={4}
                 maxRows={4}
                 multiline
@@ -143,58 +122,28 @@ const AddMileageModal = ({ semester, subitemId }: Props) => {
                 ...theme.typography.body2,
               }}
             />
-            <Flex.Row gap="1rem" align="center" wrap="wrap">
-              <UploadButton
-                label="첨부파일 업로드"
-                onUpload={file.handleChange}
-              />
+            <Flex.Row gap="1rem" align="center">
               <Flex.Column>
-                {file.value ? (
-                  <Text
-                    style={{
-                      ...theme.typography.body2,
-                      whiteSpace: 'nowrap',
-                      overflow: 'hidden',
-                      textOverflow: 'ellipsis',
-                    }}
-                  >
-                    {file.value?.name}
-                  </Text>
+                {item.file ? (
+                  <FileDownloadButton item={item} />
                 ) : (
-                  <>
-                    <Text style={{ ...theme.typography.body2 }}>
-                      활동을 증명할 파일을 업로드해주세요
-                    </Text>
-                    <Text style={{ ...theme.typography.body2 }}>
-                      첨부파일은 pdf만 업로드 가능
-                    </Text>
-                  </>
+                  <Text style={{ ...theme.typography.body2 }}>
+                    제출된 파일이 없습니다.
+                  </Text>
                 )}
               </Flex.Column>
             </Flex.Row>
             <FormField.Box />
           </FormField>
 
-          <Flex.Row justify="center" gap="1rem" margin="2rem 0 0">
-            <S.CancelButton
-              label="취소하기"
-              onClick={toggleModal}
-              size="large"
-            />
-            <S.SubmitButton
-              type="submit"
-              label="등록하기"
-              size="large"
-              onClick={() => trackAddNewMileageButton()}
-            />
+          <Flex.Row justify="center" gap="2rem" margin="2rem 0 0">
+            <S.CloseButton label="닫기" onClick={toggleModal} size="large" />
           </Flex.Row>
         </S.Form>
       </Modal.Body>
     </Modal>
   );
 };
-
-export default AddMileageModal;
 
 const S = {
   Form: styled('form')`
@@ -204,16 +153,7 @@ const S = {
     padding-top: 2rem;
     width: 100%;
   `,
-  SubmitButton: styled(Button)`
+  CloseButton: styled(Button)`
     width: 300px;
-  `,
-  CancelButton: styled(Button)`
-    background-color: ${({ theme }) => theme.palette.grey300};
-    width: 300px;
-
-    &:hover,
-    &:active {
-      background-color: ${({ theme }) => theme.palette.grey400};
-    }
   `,
 };
