+++ src/pages/AddMileagePage/components/SubmittedMileageSection.tsx
@@ -1,9 +1,11 @@
-import { Flex, SectionErrorFallback, Title } from '@/components';
-import { SubmittedMileageTable } from '@/components/AddMileage';
+import { BoxSkeleton, Flex, SectionErrorFallback, Title } from '@/components';
 import { QueryErrorResetBoundary } from '@tanstack/react-query';
+import { Suspense } from 'react';
 import { ErrorBoundary } from 'react-error-boundary';
 
-const SubmittedMileageSection = () => {
+import { SubmittedMileageTable } from './SubmittedMileageTable';
+
+export const SubmittedMileageSection = () => {
   return (
     <Flex.Column height="50%">
       <Title label="신청 완료 마일리지" />
@@ -13,12 +15,12 @@ const SubmittedMileageSection = () => {
             FallbackComponent={SectionErrorFallback}
             onReset={reset}
           >
-            <SubmittedMileageTable />
+            <Suspense fallback={<BoxSkeleton />}>
+              <SubmittedMileageTable />
+            </Suspense>
           </ErrorBoundary>
         )}
       </QueryErrorResetBoundary>
     </Flex.Column>
   );
 };
-
-export default SubmittedMileageSection;
