+++ src/pages/AddMileagePage/hooks/useGetEtcMileageQuery.ts
@@ -1,0 +1,14 @@
+import { QUERY_KEYS } from '@/constants/queryKeys';
+import { useSuspenseQuery } from '@tanstack/react-query';
+
+import { getEtcMileageList } from '../apis/addMileage';
+import { EtcMileageResponse } from '../types/addMileage';
+
+export const useGetEtcMileageQuery = () => {
+  const { data, ...rest } = useSuspenseQuery<EtcMileageResponse[]>({
+    queryKey: [QUERY_KEYS.etcMileageList],
+    queryFn: () => getEtcMileageList(),
+  });
+
+  return { etcMileageList: data, ...rest };
+};
