+++ src/pages/AddMileagePage/hooks/useGetSubmittedFileQuery.ts
@@ -1,12 +1,13 @@
-import { getSubmittedMileageFile } from '@/apis/mileage';
 import { QUERY_KEYS } from '@/constants/queryKeys';
 import { useQuery } from '@tanstack/react-query';
 
-const useGetSubmittedFileQuery = (uniqueFileName: string) => {
-  return useQuery<Blob>({
+import { getSubmittedMileageFile } from '../apis/addMileage';
+
+export const useGetSubmittedFileQuery = (uniqueFileName: string) => {
+  const { data, ...rest } = useQuery<Blob>({
     queryKey: [QUERY_KEYS.submittedMileage, uniqueFileName],
     queryFn: () => getSubmittedMileageFile({ uniqueFileName }),
   });
-};
 
-export default useGetSubmittedFileQuery;
+  return { file: data, ...rest };
+};
